requirements.txt
flask==1.1.2
datetime
uuid
Flask-SQLAlchemy
pyjwt